requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.53
langchain_google_community==2.0.7
langchain_openai==0.3.14
langgraph==0.3.31
openai==1.75.0
python-dotenv==1.1.0
